feat: openapi validation for values
- search schemas in openapi directories
- validate values on startup and ConfigMap change
- validate values before applying values patches
