+++ go.mod
@@ -4,10 +4,14 @@ go 1.12
 
 require (
 	github.com/davecgh/go-spew v1.1.1
-	github.com/evanphx/json-patch v4.5.0+incompatible
+	github.com/evanphx/json-patch v4.9.0+incompatible
 	github.com/flant/shell-operator v1.0.0-beta.13 // branch: master
 	github.com/go-chi/chi v4.0.3+incompatible
 	github.com/go-openapi/spec v0.19.3
+	github.com/go-openapi/strfmt v0.19.3
+	github.com/go-openapi/swag v0.19.5
+	github.com/go-openapi/validate v0.19.7
+	github.com/hashicorp/go-multierror v1.0.0
 	github.com/kennygrant/sanitize v1.2.4
 	github.com/onsi/ginkgo v1.11.0
 	github.com/onsi/gomega v1.9.0
